Extract reusable hooks
